requirements.txt
black
langchain
openai
pyrogram
python-dotenv
tgcrypto
transformers
uvloop
pydub